feat(core): domain-free core step 5 — tools cluster split (#886)
Step 5 of 8 in the domain-free-core refactor (see
docs/architecture/domain-free-core-audit.md §6).

Whole-file move:
- core/tools/analysis.py → plugins/game_ip/tools/analysis.py
  (~285 LOC; RunAnalystTool, RunEvaluatorTool, PSMCalculateTool,
  ExplainScoreTool — all 100% game-IP per audit §3.2 row 9)

Three-way split of core/tools/signal_tools.py (~640 LOC, file retired):
- 5 IP signal tools (YouTube/Reddit/Twitch/Steam/GoogleTrends) +
  _load_signal helper → plugins/game_ip/tools/signal_tools.py
  (~277 LOC)
- _parse_mcp_content / _try_mcp_signal (renamed parse_mcp_content /
  try_mcp_signal, public API) → core/mcp/utils.py (NEW, 81 LOC) —
  reusable MCP-fallback helpers for any future plugin signal layer
- WebSearchTool (3-provider generic, no plugin dep) →
  core/tools/web_search.py (NEW, 201 LOC) — schema inlined as
  module constant per the step-2 data_tools precedent

JSON registry split:
- core/tools/tool_schemas.json retired (only consumers were
  analysis.py + signal_tools.py, both moving)
- 9 plugin entries (4 analysis + 5 signal) → plugins/game_ip/tools/
  tool_schemas.json, loaded by both plugin modules
- WebSearchTool schema inlined in core/tools/web_search.py (no
  centralized JSON consumer, matches step-2 inline pattern)

TID251 ban entries (per docs/architecture/naming-conventions.md §3):
- core.tools.analysis → plugins.game_ip.tools.analysis (v0.66.2 step 5)
- core.tools.signal_tools → split notice across all 3 destinations
  (v0.66.2 step 5)

Caller updates (6 files): core/lifecycle/container.py (lazy plugin
imports matching step-2 precedent), plugins/game_ip/cli/tool_handlers.py
(GAME_IP_DELEGATED_TOOLS rewired), 5 test files.

Verification:
- ruff/format/mypy/deptry/codespell: clean
- mypy source-file count: 245 → 247 (+2 new core utility modules)
- pytest -m "not live": 4386 passed, 1 skipped, 24 deselected
- E2E `geode analyze "Cowboy Bebop" --dry-run`: A (68.4) unchanged
- Negative tests: TID251 fires for both old paths
