+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Versioning follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Architecture
+- **Domain-free core, step 5 of 8 (`docs/architecture/domain-free-core-audit.md`).** Tools cluster split — `core/tools/analysis.py` (whole 285-LOC file) moved to `plugins/game_ip/tools/analysis.py`; `core/tools/signal_tools.py` (640 LOC) fully retired with symbols dispersed across three destinations: the 5 IP signal scrapers (`YouTubeSearchTool`, `RedditSentimentTool`, `TwitchStatsTool`, `SteamInfoTool`, `GoogleTrendsTool`) plus the `_load_signal` fixture helper moved to new `plugins/game_ip/tools/signal_tools.py`; the reusable MCP-fallback infrastructure (`_parse_mcp_content`, `_try_mcp_signal`) was promoted to a public API surface at new `core/mcp/utils.py` (renamed without underscores: `parse_mcp_content`, `try_mcp_signal`) so any future plugin's signal layer can adopt the same MCP-first / fixture-fallback shape; the generic 3-provider (Anthropic / OpenAI / GLM) `WebSearchTool` moved to new `core/tools/web_search.py` since it has no game-IP coupling. `core/tools/tool_schemas.json` (the only consumers were `analysis.py` and `signal_tools.py`) was retired; the 9 plugin-coupled schema entries (4 analysis tools + 5 signal tools) moved to `plugins/game_ip/tools/tool_schemas.json` loaded by the plugin modules; the generic `WebSearchTool` schema is inlined as a module constant in `core/tools/web_search.py` (matching the step-2 `plugins/game_ip/tools/data_tools.py:QueryMonoLakeTool` precedent of inline schemas for plugin/generic tools). Caller updates: `core/lifecycle/container.py:build_default_registry` switches to lazy-import the analysis quartet and the 5 signal tools from `plugins.game_ip.tools.*` (matching the existing `QueryMonoLakeTool` lazy-import shape), and lazy-imports `WebSearchTool` from `core.tools.web_search`; `plugins/game_ip/cli/tool_handlers.py:GAME_IP_DELEGATED_TOOLS` rewires its 4 signal entries to `plugins.game_ip.tools.signal_tools`; `tests/test_analysis_tools.py`, `tests/test_e2e.py`, `tests/test_signal_tools.py`, `tests/test_signal_tools_mcp.py`, `tests/test_native_tools.py` updated to the new import paths (the MCP test now imports `parse_mcp_content` / `try_mcp_signal` from `core.mcp.utils` and aliases them locally to keep the test body unchanged). Two new TID251 banned-api entries land in `pyproject.toml`: `core.tools.analysis` → single-target message; `core.tools.signal_tools` → triple-destination message pointing at `plugins.game_ip.tools.signal_tools` / `core.mcp.utils` / `core.tools.web_search` so a stale import gets the full breadcrumb. After step 5, `grep -rn "core\.tools\.signal_tools\|core\.tools\.analysis" core/ tests/ plugins/` returns zero hits. Quality gates green: ruff/ruff-format/mypy/deptry/codespell clean, 4386 tests pass, E2E anchor `geode analyze "Cowboy Bebop" --dry-run` unchanged at A (68.4) undermarketed. See `docs/architecture/naming-conventions.md` §1 (path mirroring) and §3 (TID251 message format) for the conventions applied. (`plugins/game_ip/tools/{analysis,signal_tools,tool_schemas.json}`, `core/mcp/utils.py`, `core/tools/web_search.py`, `core/lifecycle/container.py`, `plugins/game_ip/cli/tool_handlers.py`, `tests/test_{analysis_tools,e2e,signal_tools,signal_tools_mcp,native_tools}.py`, `pyproject.toml`)
+
 - **Domain-free core, step 4 of 8 (`docs/architecture/domain-free-core-audit.md`).** CLI commands and tool-handler IP halves split out of `core/cli/`. `core/cli/commands.py` lost `cmd_list`, `cmd_generate`, `cmd_batch`, the 14 game-IP slash entries in `COMMAND_MAP` (`/analyze`, `/run`, `/list`, `/search`, `/report`, `/batch`, `/compare`, `/generate` + their aliases), and the IP examples block in `show_help`; the slashes now live in `plugins/game_ip/cli/commands.py:GAME_IP_SLASHES` and merge back into the generic `COMMAND_MAP` at bootstrap via the new `install_domain_commands(domain)` helper. `core/cli/tool_handlers.py` lost `_build_analysis_handlers` (180 LOC: `handle_list_ips` / `handle_analyze_ip` / `handle_search_ips` / `handle_compare_ips` / `handle_generate_report` / `handle_batch_analyze`), `handle_generate_data` (in `_build_execution_handlers`), the four signal entries in `_DELEGATED_TOOLS` (`youtube_search`, `reddit_sentiment`, `steam_info`, `google_trends`), and the `FIXTURE_MAP` reads in `handle_check_status` / `handle_show_help`; those handlers now live in `plugins/game_ip/cli/tool_handlers.py:build_game_ip_handlers()` and merge into the dispatcher dict at handler-build time via the new `install_domain_tool_handlers(handlers)` helper. The `handle_rerun_node` allowlist (`{"scoring", "verification", "synthesizer"}`) is now sourced from `domain.get_rerunnable_nodes()` instead of being hardcoded in core. `handle_check_status`'s `fixture_count` is now sourced from `domain.list_fixtures()`. `show_help` defers the IP-specific block to `domain.render_help_fragment()`. Three new optional `DomainPort` v2 methods declared in `core/domains/port.py` and implemented in `plugins/game_ip/adapter.py`: `get_rerunnable_nodes()`, `register_slash_commands(command_map)`, `register_tool_handlers(handlers)` — all use lazy plugin imports inside each method to avoid circular-import risk. `plugins/game_ip/__init__.py` also eagerly merges `GAME_IP_SLASHES` into `COMMAND_MAP` at import time so static-import paths (tests, REPL bootstrap, the legacy `COMMAND_REGISTRY` parity check) see the full slash registry without needing the bootstrap helper. After step 4, `grep -rn "from plugins\.game_ip\|import plugins\.game_ip" core/cli/{commands,tool_handlers}.py` returns zero hits except in `_handle_command` lazy imports for `/list`/`/generate`/`/batch` dispatch (which can't be removed until step 7-8 retire the `_handle_command` god method itself). `core/cli/routing.py` `/list` `handler_path` updated to `plugins.game_ip.cli.commands:cmd_list`. No TID251 entries land this step — step 4 only relocates symbols inside still-existing modules (`core/cli/commands.py` and `core/cli/tool_handlers.py` both shrink but stay), so module-level bans don't apply; resumes in step 5 when `core/tools/analysis.py` and the IP half of `core/tools/signal_tools.py` move whole-module. Quality gates green: ruff/mypy clean, deptry clean, 4386+ tests pass, E2E anchor `geode analyze "Cowboy Bebop" --dry-run` unchanged at A (68.4) undermarketed. (`core/domains/port.py`, `core/cli/{commands,tool_handlers,bootstrap,routing,__init__}.py`, `plugins/game_ip/__init__.py`, `plugins/game_ip/adapter.py`, `plugins/game_ip/cli/{commands,tool_handlers}.py`, `tests/test_commands.py`)
 
 ## [0.66.1] — 2026-05-06
